requirements.txt
asgiref==3.4.1
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-bootstrap-v5==1.0.5
django-crum==0.7.9
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-sms==0.5.0
djangorestframework==3.12.4
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==3.10.1
oauthlib==3.1.1
phonenumbers==8.12.37
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.2.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
twilio==7.3.0
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.5.0